search-api/requirements.txt
Babel==2.12.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask==2.2.3
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
SQLAlchemy==2.0.6
Werkzeug==2.2.3
alembic==1.10.2
attrs==22.2.0
blinker==1.5
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
datedelta==1.4
dpath==2.1.4
ecdsa==0.18.0
expiringdict==1.2.2
flask-babel==3.0.1
flask-jwt-oidc==0.3.0
google-api-core==2.11.0
google-auth-oauthlib==1.0.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-pubsub==2.15.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio-status==1.51.3
grpcio==1.51.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
jsonschema==3.2.0
launchdarkly-server-sdk==8.1.1
oauthlib==3.2.2
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyRFC3339==1.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2022.7.1
requests-oauthlib==1.3.1
requests==2.28.2
rsa==4.9
semver==2.13.0
sentry-sdk==1.16.0
six==1.16.0
strict-rfc3339==0.7
typing_extensions==4.5.0
urllib3==1.26.15
git+https://github.com/bcgov/registry-schemas.git@1.5.3#egg=registry_schemas
git+https://github.com/daxiom/simple-cloudevent.py.git@0.0.2
git+https://github.com/daxiom/flask-pub.git@0.0.4